requirements.txt
folium==0.14.0
nltk==3.8.1
numpy==1.24.1
pandas==1.5.3
streamlit_folium==0.11.0
translators==5.5.6
